+++ dotnetcore/QRCodeMain/Controllers/QrCodeController.cs
@@ -25,7 +25,7 @@ namespace QRCodeMain.Controllers
         }
 
         // GET: QrCode/Details/5
-        public async Task<IActionResult> Details(int? id)
+        public async Task<IActionResult> Details(Int64? id)
         {
             if (id == null)
             {
@@ -67,7 +67,7 @@ namespace QRCodeMain.Controllers
         }
 
         // GET: QrCode/Edit/5
-        public async Task<IActionResult> Edit(int? id)
+        public async Task<IActionResult> Edit(Int64? id)
         {
             if (id == null)
             {
@@ -87,7 +87,7 @@ namespace QRCodeMain.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("QrCodeId")] QrCode qrCode)
+        public async Task<IActionResult> Edit(Int64 id, [Bind("QrCodeId")] QrCode qrCode)
         {
             if (id != qrCode.QrCodeId)
             {
@@ -118,7 +118,7 @@ namespace QRCodeMain.Controllers
         }
 
         // GET: QrCode/Delete/5
-        public async Task<IActionResult> Delete(int? id)
+        public async Task<IActionResult> Delete(Int64? id)
         {
             if (id == null)
             {
@@ -138,7 +138,7 @@ namespace QRCodeMain.Controllers
         // POST: QrCode/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(int id)
+        public async Task<IActionResult> DeleteConfirmed(Int64 id)
         {
             var qrCode = await _context.QrCodes.SingleOrDefaultAsync(m => m.QrCodeId == id);
             _context.QrCodes.Remove(qrCode);
@@ -146,7 +146,7 @@ namespace QRCodeMain.Controllers
             return RedirectToAction(nameof(Index));
         }
 
-        private bool QrCodeExists(int id)
+        private bool QrCodeExists(Int64 id)
         {
             return _context.QrCodes.Any(e => e.QrCodeId == id);
         }
